# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: Agency Gateway API
  version: 0.1.0
  description: |
    REST API for the Agency AI agent operating platform. Deploys agents in
    enforced isolation containers with credential scoping, network mediation,
    read-only constraints, and continuous audit.

    All endpoints except /health require API token authentication via the
    X-Agency-Token header. Streaming endpoints accept `Accept: application/x-ndjson`.
  license:
    name: Proprietary
servers:
  - url: http://localhost:8200/api/v1
    description: Local gateway daemon
security:
  - apiToken: []
...
paths:
...
  /missions:
    post:
      operationId: createMission
      summary: Create a new mission
      tags: [Missions]
      requestBody:
        required: true
        content:
          application/x-yaml:
            schema:
              type: object
              description: Mission definition in YAML
      responses:
        "201":
          description: Mission created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"
    get:
      operationId: listMissions
      summary: List all missions
      tags: [Missions]
      responses:
        "200":
          description: Mission list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Mission"
  /missions/{name}:
    get:
      operationId: showMission
      summary: Show mission detail
      tags: [Missions]
      parameters:
        - $ref: "#/components/parameters/missionName"
      responses:
        "200":
          description: Mission detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateMission
      summary: Update a mission
      tags: [Missions]
      parameters:
        - $ref: "#/components/parameters/missionName"
      requestBody:
        required: true
        content:
          application/x-yaml:
            schema:
              type: object
              description: Updated mission definition in YAML
      responses:
        "200":
          description: Mission updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"
    delete:
      operationId: deleteMission
      summary: Delete a mission
      tags: [Missions]
      parameters:
        - $ref: "#/components/parameters/missionName"
      responses:
        "200":
          description: Mission deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
  /missions/{name}/assign:
    post:
      operationId: assignMission
      summary: Assign a mission to an agent or team
      description: |
        Returns 422 with structured pre-flight result when checks fail.
        ASK tenet 11: delegation cannot exceed delegator scope.
      tags: [Missions]
      parameters:
        - $ref: "#/components/parameters/missionName"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target]
              properties:
                target:
                  type: string
                type:
                  type: string
                  enum: [agent, team]
                  default: agent
      responses:
        "200":
          description: Mission assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "422":
          description: Pre-flight checks failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  preflight:
                    type: object
  /missions/{name}/pause:
    post:
      operationId: pauseMission
      summary: Pause a mission
      tags: [Missions]
      parameters:
        - $ref: "#/components/parameters/missionName"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        "200":
          description: Mission paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
  /missions/{name}/resume:
    post:
      operationId: resumeMission
      summary: Resume a paused mission
      tags: [Missions]
      parameters:
        - $ref: "#/components/parameters/missionName"
      responses:
        "200":
          description: Mission resumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
  /missions/{name}/complete:
    post:
      operationId: completeMission
      summary: Mark a mission as completed
      tags: [Missions]
      parameters:
        - $ref: "#/components/parameters/missionName"
      responses:
        "200":
          description: Mission completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
  /missions/{name}/history:
    get:
      operationId: missionHistory
      summary: Get mission history entries
      tags: [Missions]
      parameters:
        - $ref: "#/components/parameters/missionName"
      responses:
        "200":
          description: History entries
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
...
  /missions/{name}/knowledge:
    post:
      operationId: missionKnowledge
      summary: Query knowledge graph scoped to a mission
      description: ASK tenet 24 — knowledge access bounded by authorization scope.
      tags: [Missions]
      parameters:
        - $ref: "#/components/parameters/missionName"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
      responses:
        "200":
          description: Scoped knowledge results
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /missions/{name}/claim:
    post:
      operationId: claimMissionEvent
      summary: Claim an event for deconfliction (no-coordinator teams)
      tags: [Missions]
      parameters:
        - $ref: "#/components/parameters/missionName"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [event_key]
              properties:
                event_key:
                  type: string
                agent_name:
                  type: string
      responses:
        "200":
          description: Claim result
          content:
            application/json:
              schema:
                type: object
                properties:
                  claimed:
                    type: boolean
                  holder:
                    type: string
    delete:
      operationId: releaseMissionClaim
      summary: Release a previously claimed event
      tags: [Missions]
      parameters:
        - $ref: "#/components/parameters/missionName"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                event_key:
                  type: string
      responses:
        "200":
          description: Claim released
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"

  # ── Memory ─────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    apiToken:
      type: apiKey
      in: header
      name: X-Agency-Token
      description: API token generated during `agency init`
    webhookSecret:
      type: apiKey
      in: header
      name: X-Webhook-Secret
      description: Per-webhook secret for inbound webhook authentication
  parameters:
...
    missionName:
      name: name
      in: path
      required: true
      schema:
        type: string
      description: Mission name or ID
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    StatusResponse:
      type: object
      properties:
        status:
          type: string
        name:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
...
    Mission:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        version:
          type: integer
        status:
          type: string
        assigned_to:
          type: string
        assigned_type:
          type: string
        meeseeks:
          type: boolean
        meeseeks_limit:
          type: integer
        meeseeks_model:
          type: string
        meeseeks_budget:
          type: number
        triggers:
          type: array
          items:
            type: object
        reflection:
          $ref: "#/components/schemas/MissionReflection"
        success_criteria:
          $ref: "#/components/schemas/MissionSuccessCriteria"
        fallback:
          $ref: "#/components/schemas/MissionFallback"
        cost_mode:
          type: string
          enum: [frugal, balanced, thorough]
        min_task_tier:
          type: string
          enum: [minimal, standard, full]
        procedural_memory:
          type: object
          properties:
            capture:
              type: boolean
            retrieve:
              type: boolean
            max_retrieved:
              type: integer
            include_failures:
              type: boolean
        episodic_memory:
          type: object
          properties:
            capture:
              type: boolean
            retrieve:
              type: boolean
            max_retrieved:
              type: integer
            retention_days:
              type: integer
            tool_enabled:
              type: boolean
...
    MissionReflection:
      type: object
      properties:
        enabled:
          type: boolean
          default: false
        max_rounds:
          type: integer
          default: 3
          minimum: 1
          maximum: 10
        criteria:
          type: array
          items:
            type: string
    MissionSuccessCriteria:
      type: object
      properties:
        checklist:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              description:
                type: string
              required:
                type: boolean
        evaluation:
          type: object
          properties:
            enabled:
              type: boolean
              default: false
            mode:
              type: string
              enum: [llm, checklist_only]
            model:
              type: string
              default: default
            on_failure:
              type: string
              enum: [flag, retry, block]
              default: flag
            max_retries:
              type: integer
              default: 1
    MissionFallback:
      type: object
      properties:
        policies:
          type: array
          items:
            type: object
            properties:
              trigger:
                type: string
                enum: [tool_error, capability_unavailable, budget_warning, consecutive_errors, timeout, no_progress]
              tool:
                type: string
              capability:
                type: string
              threshold:
                type: integer
              count:
                type: integer
              strategy:
                type: array
                items:
                  type: object
                  properties:
                    action:
                      type: string
                      enum: [retry, alternative_tool, degrade, simplify, complete_partial, pause_and_assess, escalate]
                    max_attempts:
                      type: integer
                    backoff:
                      type: string
                      enum: [exponential, fixed, none]
                    tool:
                      type: string
                    hint:
                      type: string
                    severity:
                      type: string
                    message:
                      type: string
        default_policy:
          type: object
          properties:
            strategy:
              type: array
              items:
                type: object